[PRMDR-161] Extract auth policy and http verb class to separate file
