+++ lambdas/handlers/authoriser_handler.py
@@ -12,7 +12,6 @@ See the License for the specific language governing permissions and limitations 
 
 import logging
 import os
-import re
 import time
 
 import boto3
@@ -20,6 +19,7 @@ import botocore.exceptions
 import jwt
 from boto3.dynamodb.conditions import Key
 from enums.permitted_role import PermittedRole
+from models.auth_policy import AuthPolicy, HttpVerb
 from services.dynamo_service import DynamoDBService
 from utils.exceptions import AuthorisationException
 from utils.get_aws_region import get_aws_region
@@ -64,7 +64,7 @@ def lambda_handler(event, context):
 
     principal_id = ""
     _, _, _, region, aws_account_id, api_gateway_arn = event["methodArn"].split(":")
-    api_id, stage, http_verb, resource_name = api_gateway_arn.split("/")
+    api_id, stage, _http_verb, _resource_name = api_gateway_arn.split("/")
 
     policy = AuthPolicy(principal_id, aws_account_id)
     policy.restApiId = api_id
@@ -134,151 +134,3 @@ def validate_login_session(current_session, ndr_session_id):
         raise AuthorisationException(
             f"The session is already expired for session ID ending in: {redact_id(ndr_session_id)}"
         )
-
-
-class HttpVerb:
-    GET = "GET"
-    POST = "POST"
-    PUT = "PUT"
-    PATCH = "PATCH"
-    HEAD = "HEAD"
-    DELETE = "DELETE"
-    OPTIONS = "OPTIONS"
-    ALL = "*"
-
-
-class AuthPolicy(object):
-    awsAccountId = ""
-    principalId = ""
-    version = "2012-10-17"
-    pathRegex = r"^[/.a-zA-Z0-9-*]+$"
-    allowMethods = []
-    denyMethods = []
-
-    restApiId = "<<restApiId>"
-    region = get_aws_region()
-    stage = "dev"
-
-    def __init__(self, principal, awsAccountId):
-        self.awsAccountId = awsAccountId
-        self.principalId = principal
-        self.allowMethods = []
-        self.denyMethods = []
-
-    def _addMethod(self, effect, verb, resource, conditions):
-        """Adds a method to the internal lists of allowed or denied methods. Each object in
-        the internal list contains a resource ARN and a condition statement. The condition
-        statement can be null."""
-        if verb != "*" and not hasattr(HttpVerb, verb):
-            raise NameError(
-                "Invalid HTTP verb " + verb + ". Allowed verbs in HttpVerb class"
-            )
-        resourcePattern = re.compile(self.pathRegex)
-        if not resourcePattern.match(resource):
-            raise NameError(
-                "Invalid resource path: "
-                + resource
-                + ". Path should match "
-                + self.pathRegex
-            )
-
-        if resource[:1] == "/":
-            resource = resource[1:]
-
-        resourceArn = (
-            "arn:aws:execute-api:"
-            + self.region
-            + ":"
-            + self.awsAccountId
-            + ":"
-            + self.restApiId
-            + "/"
-            + self.stage
-            + "/"
-            + verb
-            + "/"
-            + resource
-        )
-
-        if effect.lower() == "allow":
-            self.allowMethods.append(
-                {"resourceArn": resourceArn, "conditions": conditions}
-            )
-        elif effect.lower() == "deny":
-            self.denyMethods.append(
-                {"resourceArn": resourceArn, "conditions": conditions}
-            )
-
-    def _getEmptyStatement(self, effect):
-        """Returns an empty statement object prepopulated with the correct action and the
-        desired effect."""
-        statement = {
-            "Action": "execute-api:Invoke",
-            "Effect": effect[:1].upper() + effect[1:].lower(),
-            "Resource": [],
-        }
-
-        return statement
-
-    def _getStatementForEffect(self, effect, methods):
-        """This function loops over an array of objects containing a resourceArn and
-        conditions statement and generates the array of statements for the policy."""
-        statements = []
-
-        if len(methods) > 0:
-            statement = self._getEmptyStatement(effect)
-
-            for curMethod in methods:
-                if curMethod["conditions"] is None or len(curMethod["conditions"]) == 0:
-                    statement["Resource"].append(curMethod["resourceArn"])
-                else:
-                    conditionalStatement = self._getEmptyStatement(effect)
-                    conditionalStatement["Resource"].append(curMethod["resourceArn"])
-                    conditionalStatement["Condition"] = curMethod["conditions"]
-                    statements.append(conditionalStatement)
-
-            statements.append(statement)
-
-        return statements
-
-    def allowAllMethods(self):
-        """Adds a '*' allow to the policy to authorize access to all methods of an API"""
-        self._addMethod("Allow", HttpVerb.ALL, "*", [])
-
-    def denyAllMethods(self):
-        """Adds a '*' allow to the policy to deny access to all methods of an API"""
-        self._addMethod("Deny", HttpVerb.ALL, "*", [])
-
-    def allowMethod(self, verb, resource):
-        """Adds an API Gateway method (Http verb + Resource path) to the list of allowed
-        methods for the policy"""
-        self._addMethod("Allow", verb, resource, [])
-
-    def denyMethod(self, verb, resource):
-        """Adds an API Gateway method (Http verb + Resource path) to the list of denied
-        methods for the policy"""
-        self._addMethod("Deny", verb, resource, [])
-
-    def build(self):
-        """Generates the policy document based on the internal lists of allowed and denied
-        conditions. This will generate a policy with two main statements for the effect:
-        one statement for Allow and one statement for Deny.
-        Methods that includes conditions will have their own statement in the policy."""
-        if (self.allowMethods is None or len(self.allowMethods) == 0) and (
-            self.denyMethods is None or len(self.denyMethods) == 0
-        ):
-            raise NameError("No statements defined for the policy")
-
-        policy = {
-            "principalId": self.principalId,
-            "policyDocument": {"Version": self.version, "Statement": []},
-        }
-
-        policy["policyDocument"]["Statement"].extend(
-            self._getStatementForEffect("Allow", self.allowMethods)
-        )
-        policy["policyDocument"]["Statement"].extend(
-            self._getStatementForEffect("Deny", self.denyMethods)
-        )
-
-        return policy
